+++ lib/user_agent/browsers.rb
@@ -24,6 +24,7 @@ require 'user_agent/browsers/facebook'
 require 'user_agent/browsers/gecko'
 require 'user_agent/browsers/google_home'
 require 'user_agent/browsers/google_podcasts'
+require 'user_agent/browsers/googlebot'
 require 'user_agent/browsers/iheart_radio'
 require 'user_agent/browsers/internet_explorer'
 require 'user_agent/browsers/itunes'
@@ -66,6 +67,7 @@ class UserAgent
       WechatBrowser,
       Vivaldi,
       Facebook,
+      Googlebot,
       GooglePodcasts,
       GoogleHome,
       Acast,