+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 # 1.3.0
-* Add browser support for Deezer
+* Add browser support for Deezer, Googlebot
 
 # 1.2.0
 * Add browsers support for Audible, BlackBerry, Podbean, ThePodcastApp, TuneIn, Wondery, general Android